# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "梭子 - 智能日程管理平台 API",
    "version": "1.0.0",
    "description": "梭子（Shuttle）是一个智能日程记录与管理平台，提供高效记录、AI辅助规划、积分激励与个性化装扮等功能。\n\n## 微服务架构\n- **user-service**: 用户注册登录、信息管理、积分账户\n- **schedule-service**: 日程CRUD、计时控制、日历视图、统计接口\n- **ai-service**: AI规划生成、每日/月度总结\n- **notification-service**: 消息推送、提醒生成\n- **mall-service**: 虚拟商品管理、购买记录、用户装扮\n\n## 认证方式\n除登录注册外，所有接口需在 Header 中携带 JWT Token：\n```\nAuthorization: Bearer <token>\n```",
    "contact": {
      "name": "梭子开发团队"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Gateway 统一入口 (开发环境)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Token 认证，登录接口返回"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码，200 表示成功",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "操作成功"
          },
          "data": {
            "description": "响应数据"
          }
        }
      },
...
      "AIPlanRequest": {
        "type": "object",
        "required": [
          "userPrompt"
        ],
        "properties": {
          "userPrompt": {
            "type": "string",
            "maxLength": 2000,
            "description": "用户的自然语言描述",
            "example": "明天上午学习Java 2小时，下午运动1小时，晚上看书"
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "description": "目标日期（默认明天）"
          },
          "preferences": {
            "type": "object",
            "description": "用户偏好设置",
            "properties": {
              "preferredStartHour": {
                "type": "integer",
                "example": 9
              },
              "preferredEndHour": {
                "type": "integer",
                "example": 22
              },
              "defaultDuration": {
                "type": "integer",
                "example": 60
              }
            }
          }
        }
      },
...
      "AISummaryRequest": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "指定日期（默认今天）"
          },
          "stream": {
            "type": "boolean",
            "description": "是否流式返回",
            "default": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/ai/plan": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "AI 生成日程规划",
        "description": "输入自然语言描述，AI 解析并生成结构化日程建议，支持流式输出（SSE）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "generatePlan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIPlanRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "stream",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "是否使用 SSE 流式返回"
          }
        ],
        "responses": {
          "200": {
            "description": "规划生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AIPlanResponse"
                        }
                      }
                    }
                  ]
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "SSE 事件流，逐条返回建议"
                }
              }
            }
          }
        }
      }
    },
    "/ai/plan/{planId}/adopt": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "采纳 AI 规划",
        "description": "一键将AI规划建议创建为实际日程",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "adoptPlan",
        "parameters": [
          {
            "name": "planId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptPlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "采纳成功，返回创建的日程列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Schedule"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ai/summary/daily": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "生成每日总结",
        "description": "对当日已完成日程进行智能总结，分析效率和时间分配。支持 SSE 流式返回。",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "generateDailySummary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISummaryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "总结生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AIReportInfo"
                        }
                      }
                    }
                  ]
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ai/summary/monthly": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "生成月度总结",
        "description": "异步生成月度总结报告，返回报告ID供后续查询",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "generateMonthlySummary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "year": {
                    "type": "integer"
                  },
                  "month": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "已提交异步生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "reportId": {
                              "type": "integer",
                              "format": "int64"
                            },
                            "status": {
                              "type": "string",
                              "enum": [
                                "GENERATING"
                              ]
                            },
                            "estimatedSeconds": {
                              "type": "integer",
                              "description": "预计生成时间（秒）"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ai/summary/yearly": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "生成年度总结",
        "description": "异步生成年度总结报告",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "generateYearlySummary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "year": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "已提交异步生成任务"
          }
        }
      }
    },
    "/ai/reports": {
      "get": {
        "tags": [
          "AI助手"
        ],
        "summary": "查询历史报告列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "listReports",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DAILY",
                "WEEKLY",
                "MONTHLY",
                "YEARLY"
              ]
            },
            "description": "报告类型"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResult"
                }
              }
            }
          }
        }
      }
    },
    "/ai/reports/{id}": {
      "get": {
        "tags": [
          "AI助手"
        ],
        "summary": "获取报告详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "getReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AIReportInfo"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AI助手"
        ],
        "summary": "删除织史",
        "description": "删除指定织史（软删除，写入 deleted_at）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "deleteReport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "404": {
            "description": "织史不存在或无权删除"
          }
        }
      }
    },
    "/ai/chat": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "AI 对话助手",
        "description": "与 AI 自由对话，获取日程管理建议，支持流式返回",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "chatWithAI",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "message"
                ],
                "properties": {
                  "message": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "conversationId": {
                    "type": "string",
                    "description": "对话ID（可选，多轮对话时传入）"
                  },
                  "stream": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI 回复",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "conversationId": {
                              "type": "string"
                            },
                            "reply": {
                              "type": "string"
                            },
                            "suggestedActions": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "type": {
                                    "type": "string",
                                    "enum": [
                                      "CREATE_SCHEDULE",
                                      "VIEW_STATS",
                                      "GENERATE_SUMMARY"
                                    ]
                                  },
                                  "label": {
                                    "type": "string"
                                  },
                                  "payload": {
                                    "type": "object"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/ai/reports/batch-delete": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "批量删除织史",
        "description": "批量软删除，事务保证全部成功或全部失败",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "batchDeleteReports",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "400": {
            "description": "ids 为空或超过 200 个"
          },
          "404": {
            "description": "部分织史不存在或无权删除，已整体回滚"
          }
        }
      }
    },
    "/ai/conversations": {
      "get": {
        "tags": [
          "AI助手"
        ],
        "summary": "会话列表",
        "description": "当前用户的 AI 会话列表，按 updated_at 倒序，分页返回",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "listConversations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PageResult"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "新建会话",
        "description": "创建新会话（超过 50 个时软删最旧）；返回完整会话",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "createConversation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ConversationVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ai/conversations/{id}": {
      "get": {
        "tags": [
          "AI助手"
        ],
        "summary": "会话详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "getConversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ConversationVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AI助手"
        ],
        "summary": "保存会话（消息/草稿/勾选/规划快照）",
        "description": "全量快照保存；单会话消息上限 100（保留最近），单条正文 5000 字符截断",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "saveConversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationSaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "保存成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ConversationVO"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AI助手"
        ],
        "summary": "删除会话（软删除）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "deleteConversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    },
    "/ai/conversations/batch-delete": {
      "post": {
        "tags": [
          "AI助手"
        ],
        "summary": "批量删除会话",
        "description": "批量软删除，事务保证全部成功或全部失败",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "batchDeleteConversations",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功"
          }
        }
      }
    }
  }
}